fix: adapt the landing page to the new project structure
