+++ src/index.js
@@ -2,19 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import './index.css';
-import Landing from './landing_page/landing_page';
-// import './ComponentSingInAndSingUp.css';
-
-// import AppBtn from './ComponentSingInAndSingUp/AppBtn'
 
 import App from './AppStructureAndRouter/App'
 
 ReactDOM.render(<App />, document.getElementById('root'));
-
-
-// ReactDOM.render(
-//   <React.StrictMode>
-//     <Landing />
-//   </React.StrictMode>,
-//   document.getElementById('root')
-// );
